set service linker module as submodule of spring deployment.AB#2048737
