+++ azure-toolkit-libs/azure-toolkit-springcloud-lib/src/main/java/com/microsoft/azure/toolkit/lib/springcloud/SpringCloudApp.java
@@ -9,8 +9,6 @@ import com.azure.resourcemanager.appplatform.models.PersistentDisk;
 import com.azure.resourcemanager.appplatform.models.SpringApp;
 import com.microsoft.azure.toolkit.lib.common.model.*;
 import com.microsoft.azure.toolkit.lib.common.operation.AzureOperation;
-import com.microsoft.azure.toolkit.lib.servicelinker.ServiceLinkerModule;
-import com.microsoft.azure.toolkit.lib.servicelinker.Consumer;
 import com.microsoft.azure.toolkit.lib.springcloud.model.SpringCloudPersistentDisk;
 import lombok.Getter;
 import org.apache.commons.lang3.StringUtils;
@@ -21,18 +19,16 @@ import java.util.*;
 
 @Getter
 public class SpringCloudApp extends AbstractAzResource<SpringCloudApp, SpringCloudCluster, SpringApp>
-    implements Startable, Deletable, Consumer {
+    implements Startable, Deletable {
 
     @Nonnull
     private final SpringCloudDeploymentModule deploymentModule;
-    private final ServiceLinkerModule linkerModule;
     @Nullable
     private SpringCloudDeployment activeDeployment = null;
 
     protected SpringCloudApp(@Nonnull String name, @Nonnull SpringCloudAppModule module) {
         super(name, module);
         this.deploymentModule = new SpringCloudDeploymentModule(this);
-        this.linkerModule = new ServiceLinkerModule(getId(), this);
     }
 
     /**
@@ -42,13 +38,11 @@ public class SpringCloudApp extends AbstractAzResource<SpringCloudApp, SpringClo
         super(origin);
         this.deploymentModule = origin.deploymentModule;
         this.activeDeployment = origin.activeDeployment;
-        this.linkerModule = origin.linkerModule;
     }
 
     protected SpringCloudApp(@Nonnull SpringApp remote, @Nonnull SpringCloudAppModule module) {
         super(remote.name(), module);
         this.deploymentModule = new SpringCloudDeploymentModule(this);
-        this.linkerModule = new ServiceLinkerModule(getId(), this);
     }
 
     @Override
@@ -67,7 +61,7 @@ public class SpringCloudApp extends AbstractAzResource<SpringCloudApp, SpringClo
     @Nonnull
     @Override
     public List<AbstractAzResourceModule<?, ?, ?>> getSubModules() {
-        return Arrays.asList(deploymentModule, linkerModule);
+        return Collections.singletonList(deploymentModule);
     }
 
     @Nonnull
@@ -162,10 +156,5 @@ public class SpringCloudApp extends AbstractAzResource<SpringCloudApp, SpringClo
 
     public boolean isPersistentDiskEnabled() {
         return Objects.nonNull(this.getPersistentDisk());
-    }
-
-    @Override
-    public ServiceLinkerModule getServiceLinkerModule() {
-        return linkerModule;
     }
 }
